WIP: during memo
