+++ luda-editor/server/server-core/src/apis/memo/list_sequence_memos.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::list_sequence_memos::{Error, Request, Response};
 
 pub async fn list_sequence_memos(
     session: Option<SessionDocument>,
-    req: rpc::list_sequence_memos::Request,
+    Request {}: Request,
 ) -> rpc::list_sequence_memos::Result {
     todo!()
 }
