+++ luda-editor/server/server-core/src/apis/project/edit_user_acl.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::edit_user_acl::{Error, Request, Response};
 
 pub async fn edit_user_acl(
     session: Option<SessionDocument>,
-    req: rpc::edit_user_acl::Request,
+    Request {}: Request,
 ) -> rpc::edit_user_acl::Result {
     todo!()
 }
