+++ luda-editor/server/server-core/src/apis/project/update_server_project_shared_data.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::update_server_project_shared_data::{Error, Request, Response};
 
 pub async fn update_server_project_shared_data(
     session: Option<SessionDocument>,
-    req: rpc::update_server_project_shared_data::Request,
+    Request {}: Request,
 ) -> rpc::update_server_project_shared_data::Result {
     todo!()
 }
