+++ luda-editor/server/server-core/src/apis/sequence/create_sequence.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::create_sequence::{Error, Request, Response};
 
 pub async fn create_sequence(
     session: Option<SessionDocument>,
-    req: rpc::create_sequence::Request,
+    Request {}: Request,
 ) -> rpc::create_sequence::Result {
     todo!()
 }
