+++ luda-editor/server/server-core/src/apis/sequence/get_sequence_and_project_shared_data.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::get_sequence_and_project_shared_data::{Error, Request, Response};
 
 pub async fn get_sequence_and_project_shared_data(
     session: Option<SessionDocument>,
-    req: rpc::get_sequence_and_project_shared_data::Request,
+    Request {}: Request,
 ) -> rpc::get_sequence_and_project_shared_data::Result {
     todo!()
 }
