+++ luda-editor/server/server-core/src/apis/sequence/list_project_sequences.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::list_project_sequences::{Error, Request, Response};
 
 pub async fn list_project_sequences(
     session: Option<SessionDocument>,
-    req: rpc::list_project_sequences::Request,
+    Request {}: Request,
 ) -> rpc::list_project_sequences::Result {
     todo!()
 }
