+++ luda-editor/server/server-core/src/apis/sequence/redo_update.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::redo_update::{Error, Request, Response};
 
 pub async fn redo_update(
     session: Option<SessionDocument>,
-    req: rpc::redo_update::Request,
+    Request {}: Request,
 ) -> rpc::redo_update::Result {
     todo!()
 }
