+++ luda-editor/server/server-core/src/apis/sequence/undo_update.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::undo_update::{Error, Request, Response};
 
 pub async fn undo_update(
     session: Option<SessionDocument>,
-    req: rpc::undo_update::Request,
+    Request {}: Request,
 ) -> rpc::undo_update::Result {
     todo!()
 }
