+++ luda-editor/server/server-core/src/apis/sequence/update_sequence.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::update_sequence::{Error, Request, Response};
 
 pub async fn update_sequence(
     session: Option<SessionDocument>,
-    req: rpc::update_sequence::Request,
+    Request {}: Request,
 ) -> rpc::update_sequence::Result {
     todo!()
 }
