+++ luda-editor/server/server-core/src/apis/sequence/update_sequence_cut.rs
@@ -1,8 +1,10 @@
+use super::shared::*;
 use crate::documents::*;
+use rpc::update_sequence_cut::{Error, Request, Response};
 
 pub async fn update_sequence_cut(
     session: Option<SessionDocument>,
-    req: rpc::update_sequence_cut::Request,
+    Request {}: Request,
 ) -> rpc::update_sequence_cut::Result {
     todo!()
 }
