+++ luda-editor/server/server-core/src/apis2/memo/mod.rs
@@ -46,54 +46,7 @@ impl rpc::MemoService<SessionDocument> for MemoService {
         req: rpc::create_memo::Request,
     ) -> std::pin::Pin<Box<dyn 'a + std::future::Future<Output = rpc::create_memo::Result> + Send>>
     {
-        Box::pin(async move {
-            if session.is_none() {
-                return Err(rpc::create_memo::Error::Unauthorized);
-            }
-            let session = session.unwrap();
-
-            let sequence_document = SequenceIndexDocumentGet {
-                pk_id: req.sequence_id,
-            }
-            .run()
-            .await
-            .map_err(|error| rpc::create_memo::Error::Unknown(error.to_string()))?;
-
-            let is_project_editor = crate::services()
-                .project_service
-                .is_project_editor(session.user_id, sequence_document.project_id)
-                .await
-                .map_err(|error| rpc::create_memo::Error::Unknown(error.to_string()))?;
-
-            if !is_project_editor {
-                return Err(rpc::create_memo::Error::Unauthorized);
-            }
-
-            let user_document = UserDocumentGet {
-                pk_id: session.user_id,
-            }
-            .run()
-            .await
-            .map_err(|error| rpc::create_memo::Error::Unknown(error.to_string()))?;
-
-            let memo_document = MemoDocument {
-                sequence_id: sequence_document.id,
-                cut_id: req.cut_id,
-                memo_id: rpc::Uuid::new_v4(),
-                content: req.content,
-                user_id: session.user_id,
-                user_name: user_document.name,
-            };
-
-            crate::dynamo_db()
-                .create_item(memo_document.clone())
-                .await
-                .map_err(|error| rpc::create_memo::Error::Unknown(error.to_string()))?;
-
-            Ok(rpc::create_memo::Response {
-                memo: memo_document.into(),
-            })
-        })
+        Box::pin(async move {})
     }
 
     fn delete_memo<'a>(
@@ -102,50 +55,6 @@ impl rpc::MemoService<SessionDocument> for MemoService {
         req: rpc::delete_memo::Request,
     ) -> std::pin::Pin<Box<dyn 'a + std::future::Future<Output = rpc::delete_memo::Result> + Send>>
     {
-        Box::pin(async move {
-            if session.is_none() {
-                return Err(rpc::delete_memo::Error::Unauthorized);
-            }
-            let session = session.unwrap();
-
-            let sequence_document = SequenceIndexDocumentGet {
-                pk_id: req.sequence_id,
-            }
-            .run()
-            .await
-            .map_err(|error| rpc::delete_memo::Error::Unknown(error.to_string()))?;
-
-            let is_project_editor = crate::services()
-                .project_service
-                .is_project_editor(session.user_id, sequence_document.project_id)
-                .await
-                .map_err(|error| rpc::delete_memo::Error::Unknown(error.to_string()))?;
-
-            if !is_project_editor {
-                return Err(rpc::delete_memo::Error::Unauthorized);
-            }
-
-            let memo_document = MemoDocumentGet {
-                pk_sequence_id: sequence_document.id,
-                sk_memo_id: req.memo_id,
-            }
-            .run()
-            .await
-            .map_err(|error| rpc::delete_memo::Error::Unknown(error.to_string()))?;
-
-            if session.user_id != memo_document.user_id {
-                return Err(rpc::delete_memo::Error::Forbidden);
-            }
-
-            MemoDocumentDelete {
-                pk_sequence_id: sequence_document.id,
-                sk_memo_id: req.memo_id,
-            }
-            .run()
-            .await
-            .map_err(|error| rpc::delete_memo::Error::Unknown(error.to_string()))?;
-
-            Ok(rpc::delete_memo::Response {})
-        })
+        Box::pin(async move {})
     }
 }
