+++ luda-editor/server/server-core/src/documents/mod.rs
@@ -1,6 +1,12 @@
+mod memo;
+mod project_image;
+mod sequence;
 mod session;
 mod user;
 
+pub use memo::*;
+pub use project_image::*;
+pub use sequence::*;
 pub use session::*;
 pub use user::*;
 
@@ -13,4 +19,48 @@ pub struct IdentityDocument {
     #[pk]
     pub id: String,
     pub user_id: rpc::Uuid,
+}
+
+#[document_macro::document]
+pub struct ProjectDocument {
+    #[pk]
+    pub id: rpc::Uuid,
+    pub owner_id: rpc::Uuid,
+    pub name: String,
+    pub shared_data_json: String,
+}
+
+#[document_macro::document]
+pub struct OwnerProjectDocument {
+    #[pk]
+    pub owner_id: rpc::Uuid,
+    #[sk]
+    pub project_id: rpc::Uuid,
+}
+
+#[document_macro::document]
+pub struct UserInProjectAclDocument {
+    #[pk]
+    pub project_id: rpc::Uuid,
+    #[sk]
+    pub user_id: rpc::Uuid,
+    pub permission: rpc::types::ProjectAclUserPermission,
+}
+
+#[document_macro::document]
+pub struct ProjectAclUserInDocument {
+    #[pk]
+    pub user_id: rpc::Uuid,
+    #[sk]
+    pub project_id: rpc::Uuid,
+    pub permission: rpc::types::ProjectAclUserPermission,
+}
+
+#[document_macro::document]
+pub struct CgDocument {
+    #[pk]
+    pub project_id: rpc::Uuid,
+    #[sk]
+    pub cg_id: rpc::Uuid,
+    pub cg_file: rpc::data::CgFile,
 }
